Move SourceOrdinaryMethodSymbol functionality that is not syntax specific into a new bas class.
This work is related to #45296.
